requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
